+++ src/Symfony/Component/Form/Extension/Core/Type/MoneyType.php
@@ -16,6 +16,7 @@ use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\Extension\Core\DataTransformer\MoneyToLocalizedStringTransformer;
 use Symfony\Component\Form\FormView;
+use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class MoneyType extends AbstractType
@@ -29,7 +30,7 @@ class MoneyType extends AbstractType
     {
         $builder
             ->addViewTransformer(new MoneyToLocalizedStringTransformer(
-                $options['precision'],
+                $options['scale'],
                 $options['grouping'],
                 null,
                 $options['divisor']
@@ -50,13 +51,27 @@ class MoneyType extends AbstractType
      */
     public function configureOptions(OptionsResolver $resolver)
     {
+        $scale = function (Options $options) {
+            if (null !== $options['precision']) {
+                trigger_error('The form option "precision" is deprecated since version 2.7 and will be removed in 3.0. Use "scale" instead.', E_USER_DEPRECATED);
+
+                return $options['precision'];
+            }
+
+            return 2;
+        };
+
         $resolver->setDefaults(array(
-            'precision' => 2,
+            // deprecated as of Symfony 2.7, to be removed in Symfony 3.0
+            'precision' => null,
+            'scale' => $scale,
             'grouping' => false,
             'divisor' => 1,
             'currency' => 'EUR',
             'compound' => false,
         ));
+
+        $resolver->setAllowedTypes('scale', 'int');
     }
 
     /**